[NEW] Detect loop considering more skills.
Previously a node is considered a loop (and stops immedidately) if it's
visited. This does not work in a scenario when player gets a new skill
in later stage and then back to previous stages for another path.

This fix now considers a loop when both conditions below are met:

1. Current stage is visited.
2. All next-stage nodes of current stages are also visited.

For #2, if there's a next-stage node not visited, it's not considered as
a loop.